총 및 총알 추가

+++ Assets/Player/PlayerStat.cs
@@ -8,7 +8,7 @@ public class PlayerStat : MonoBehaviour
     {
         public int Lv;
         public float Hp;
-        public int Stamina; // 공복, 이동
+        public float Stamina; // 공복, 이동 -> shield point
         public int STR; // 공격력
         public int DEF; // 방여력
         public int EXP; // 경험치
@@ -21,6 +21,7 @@ public class PlayerStat : MonoBehaviour
 
     public PS ps;
 
+    // 이후 사용
     public void StatDown()
     {
         var hpmp = GameObject.Find("player_State").GetComponent<StatsManager>();
